pip_requirements.txt
amqp==2.1.4
anyjson==0.3.3
appdirs==1.4.0
asgi-redis==1.0.0
asgiref==1.0.0
autobahn==0.17.1
billiard==3.5.0.2
celery==4.0.2
channels==1.0.2
constantly==15.1.0
coverage==4.3.4
daphne==1.0.1
Django==1.10.5
django-cas-ng>=3.5.6
django-webpack-loader==0.4.1
djangorestframework==3.5.3
incremental==16.10.1
kombu==4.0.2
mod-wsgi-httpd==2.4.23.1
msgpack-python==0.4.8
packaging==16.8
pamela>=0.3.0
PyMySQL==0.7.11
pyparsing==2.1.10
pytz==2016.10
redis==2.10.5
requests>=2.13.0
six==1.10.0
social-auth-app-django>=1.1.0
social-auth-core>=1.2.0
Twisted==16.6.0
txaio==2.6.0
uWSGI==2.0.14
vine==1.1.3
zope.interface==4.3.3
mysqlclient==1.3.10
